Fixed and balanced the spall system
Me and Cheezus have been testing spall changes on my server I've done to the ACE code for the past 3 days. I'd like you to review the code and or to push the modifications to the master branch or atleast the dev for future testing. I have currently only modified the spall functions and only tweaked two material data values in each armor type. The thing is the current spall system had lots of problems in how it was being calculated.

**By this I mean some of the problems were:**
1. Spall did not lose penetration over distance nor impacting armor of any type
2. Hesh spall was inconistent and heavily underpowered when compared to kinetic spall
3. In the spall calculations there was alot of proprietary and non-dynamic bullshit along with some values returning abnormally high values and or inf values when debugging
4. Regarding #3 there were also properitary constants which definitely were not any sort of unit conversion constants. Example being a constant Weight factor of 15 which comments said the weight factor meant that each spall flake was fucking 20mm spall

> (Spall was essentially 0/inf mass giant metal flakes that went mach 10 and did not lose energy, and infact penetration was being multiplied each impact)

**Some of the changes I've made include:**
1. Revamping the calculations
2. Noticed there were spall resist and armor material data variables that were not being used so I implemented them
3. Regarding #2, besides changing `sv_acfdamage.lua`, the only other scripts I've changed as I've said was the material scripts but only to slightly raise/lower `Material.spallarmor` and `Material.spallresist` (this is so that material types had more influence in spall along with affecting spall weight.)

So now the main difference between before and now is not just that the calculations were fixed and improved, but that the system is more dynamic and balanced. (Look back upon the problems listed and think about how tanks were being instantly vaporized by 75mm apfsds)