Fix displayed horsepower/torque values not being accurate
This removes the awful 1.5x torque multiplier that has been in the gearbox math since ACF2, which made all engines more powerful than their stats would imply. To compensate for the removal of the 1.5x multiplier, all engines have been buffed by 1.5x, so vehicles will perform the same. This change will only affect the value displayed in the menu and in the overlay.

Fuel usage was also decreased by 1.5x to account for this change, it should remain the same as before.